+++ src/routes/__init__.py
@@ -1,0 +1,11 @@
+from .utils import login_required, club_access_required, login, logout
+from .auth import auth_route
+from .user import user_route
+from .book import book_route
+from .club import club_route
+from .den import den_route
+from .forum import forum_route
+
+
+
+__all__=["auth_route", "user_route", "book_route", "club_route", "den_route", "forum_route"]
